Moves FileList functions to their own file.
This also creates a protected FileData API, which includes methods that
originally existed as static functions within `filedata/filedata.cc`.  Those
protected methods are now accessible to FileData and its child classes, without
being made accessible to code outside of the FileData class.
