+++ src/filedata/filedata.cc
@@ -68,10 +68,6 @@ static void file_data_check_sidecars(const GList *basename_list);
 static void file_data_disconnect_sidecar_file(FileData *target, FileData *sfd);
 
 
-static SortType filelist_sort_method = SORT_NONE;
-static gboolean filelist_sort_ascend = TRUE;
-static gboolean filelist_sort_case = TRUE;
-
 /*
  *-----------------------------------------------------------------------------
  * text conversion utils
@@ -383,7 +379,7 @@ static void file_data_set_path(FileData *fd, const gchar *path)
  *-----------------------------------------------------------------------------
  */
 
-static FileData *file_data_new(const gchar *path_utf8, struct stat *st, gboolean disable_sidecars)
+FileData *FileData::file_data_new(const gchar *path_utf8, struct stat *st, gboolean disable_sidecars)
 {
 	FileData *fd;
 	struct passwd *user;
@@ -399,7 +395,7 @@ static FileData *file_data_new(const gchar *path_utf8, struct stat *st, gboolean
 	fd = static_cast<FileData *>(g_hash_table_lookup(file_data_pool, path_utf8));
 	if (fd)
 		{
-		file_data_ref(fd);
+		::file_data_ref(fd);
 		}
 
 	if (!fd && file_data_planned_change_hash)
@@ -410,8 +406,8 @@ static FileData *file_data_new(const gchar *path_utf8, struct stat *st, gboolean
 			DEBUG_1("planned change: using %s -> %s", path_utf8, fd->path);
 			if (!isfile(fd->path))
 				{
-				file_data_ref(fd);
-				file_data_apply_ci(fd);
+				::file_data_ref(fd);
+				::file_data_apply_ci(fd);
 				}
 			else
 				{
@@ -422,7 +418,7 @@ static FileData *file_data_new(const gchar *path_utf8, struct stat *st, gboolean
 
 	if (fd)
 		{
-		if (disable_sidecars) file_data_disable_grouping(fd, TRUE);
+		if (disable_sidecars) ::file_data_disable_grouping(fd, TRUE);
 
 #ifdef DEBUG_FILEDATA
 		gboolean changed =
@@ -481,7 +477,7 @@ static FileData *file_data_new(const gchar *path_utf8, struct stat *st, gboolean
 	return fd;
 }
 
-static FileData *file_data_new_local(const gchar *path, struct stat *st, gboolean disable_sidecars)
+FileData *FileData::file_data_new_local(const gchar *path, struct stat *st, gboolean disable_sidecars)
 {
 	gchar *path_utf8 = path_to_utf8(path);
 	FileData *ret = file_data_new(path_utf8, st, disable_sidecars);
@@ -1164,124 +1160,6 @@ void FileData::file_data_disable_grouping_list(GList *fd_list, gboolean disable)
 }
 
 
-
-/*
- *-----------------------------------------------------------------------------
- * filelist sorting
- *-----------------------------------------------------------------------------
- */
-
-
-gint FileData::filelist_sort_compare_filedata(FileData *fa, FileData *fb)
-{
-	gint ret;
-	if (!filelist_sort_ascend)
-		{
-		std::swap(fa, fb);
-		}
-
-	switch (filelist_sort_method)
-		{
-		case SORT_NAME:
-			break;
-		case SORT_SIZE:
-			if (fa->size < fb->size) return -1;
-			if (fa->size > fb->size) return 1;
-			/* fall back to name */
-			break;
-		case SORT_TIME:
-			if (fa->date < fb->date) return -1;
-			if (fa->date > fb->date) return 1;
-			/* fall back to name */
-			break;
-		case SORT_CTIME:
-			if (fa->cdate < fb->cdate) return -1;
-			if (fa->cdate > fb->cdate) return 1;
-			/* fall back to name */
-			break;
-		case SORT_EXIFTIME:
-			if (fa->exifdate < fb->exifdate) return -1;
-			if (fa->exifdate > fb->exifdate) return 1;
-			/* fall back to name */
-			break;
-		case SORT_EXIFTIMEDIGITIZED:
-			if (fa->exifdate_digitized < fb->exifdate_digitized) return -1;
-			if (fa->exifdate_digitized > fb->exifdate_digitized) return 1;
-			/* fall back to name */
-			break;
-		case SORT_RATING:
-			if (fa->rating < fb->rating) return -1;
-			if (fa->rating > fb->rating) return 1;
-			/* fall back to name */
-			break;
-		case SORT_CLASS:
-			if (fa->format_class < fb->format_class) return -1;
-			if (fa->format_class > fb->format_class) return 1;
-			/* fall back to name */
-			break;
-		case SORT_NUMBER:
-			ret = strcmp(fa->collate_key_name_natural, fb->collate_key_name_natural);
-			if (ret != 0) return ret;
-			/* fall back to name */
-			break;
-		default:
-			break;
-		}
-
-	if (filelist_sort_case)
-		ret = strcmp(fa->collate_key_name, fb->collate_key_name);
-	else
-		ret = strcmp(fa->collate_key_name_nocase, fb->collate_key_name_nocase);
-
-	if (ret != 0) return ret;
-
-	/* do not return 0 unless the files are really the same
-	   file_data_pool ensures that original_path is unique
-	*/
-	return strcmp(fa->original_path, fb->original_path);
-}
-
-gint FileData::filelist_sort_compare_filedata_full(FileData *fa, FileData *fb, SortType method, gboolean ascend)
-{
-	filelist_sort_method = method;
-	filelist_sort_ascend = ascend;
-	return filelist_sort_compare_filedata(fa, fb);
-}
-
-static gint filelist_sort_file_cb(gpointer a, gpointer b)
-{
-	return filelist_sort_compare_filedata(static_cast<FileData *>(a), static_cast<FileData *>(b));
-}
-
-GList *FileData::filelist_sort_full(GList *list, SortType method, gboolean ascend, gboolean case_sensitive, GCompareFunc cb)
-{
-	filelist_sort_method = method;
-	filelist_sort_ascend = ascend;
-	filelist_sort_case = case_sensitive;
-	return g_list_sort(list, cb);
-}
-
-GList *FileData::filelist_insert_sort_full(GList *list, gpointer data, SortType method, gboolean ascend, gboolean case_sensitive, GCompareFunc cb)
-{
-	filelist_sort_method = method;
-	filelist_sort_ascend = ascend;
-	filelist_sort_case = case_sensitive;
-	return g_list_insert_sorted(list, data, cb);
-}
-
-GList *FileData::filelist_sort(GList *list, SortType method, gboolean ascend, gboolean case_sensitive)
-{
-	return filelist_sort_full(list, method, ascend, case_sensitive, reinterpret_cast<GCompareFunc>(filelist_sort_file_cb));
-}
-
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wunused-function"
-GList *FileData::filelist_insert_sort_unused(GList *list, FileData *fd, SortType method, gboolean ascend)
-{
-	return filelist_insert_sort_full(list, fd, method, ascend, ascend, (GCompareFunc) filelist_sort_file_cb);
-}
-#pragma GCC diagnostic pop
-
 /*
  *-----------------------------------------------------------------------------
  * basename hash - grouping of sidecars in filelist
@@ -1289,12 +1167,12 @@ GList *FileData::filelist_insert_sort_unused(GList *list, FileData *fd, SortType
  */
 
 
-static GHashTable *file_data_basename_hash_new()
+GHashTable *FileData::file_data_basename_hash_new()
 {
 	return g_hash_table_new_full(g_str_hash, g_str_equal, g_free, nullptr);
 }
 
-static GList * file_data_basename_hash_insert(GHashTable *basename_hash, FileData *fd)
+GList *FileData::file_data_basename_hash_insert(GHashTable *basename_hash, FileData *fd)
 {
 	GList *list;
 	gchar *basename = g_strndup(fd->path, fd->extension - fd->path);
@@ -1334,7 +1212,7 @@ static GList * file_data_basename_hash_insert(GHashTable *basename_hash, FileDat
 
 	if (!g_list_find(list, fd))
 		{
-		list = g_list_insert_sorted(list, file_data_ref(fd), file_data_sort_by_ext);
+		list = g_list_insert_sorted(list, ::file_data_ref(fd), file_data_sort_by_ext);
 		g_hash_table_insert(basename_hash, basename, list);
 		}
 	else
@@ -1344,183 +1222,28 @@ static GList * file_data_basename_hash_insert(GHashTable *basename_hash, FileDat
 	return list;
 }
 
-static void file_data_basename_hash_insert_cb(gpointer fd, gpointer basename_hash)
+void FileData::file_data_basename_hash_insert_cb(gpointer fd, gpointer basename_hash)
 {
 	file_data_basename_hash_insert(static_cast<GHashTable *>(basename_hash), static_cast<FileData *>(fd));
 }
 
-static void file_data_basename_hash_remove_list(gpointer, gpointer value, gpointer)
+void FileData::file_data_basename_hash_remove_list(gpointer, gpointer value, gpointer)
 {
 	filelist_free(static_cast<GList *>(value));
 }
 
-static void file_data_basename_hash_free(GHashTable *basename_hash)
+void FileData::file_data_basename_hash_free(GHashTable *basename_hash)
 {
 	g_hash_table_foreach(basename_hash, file_data_basename_hash_remove_list, nullptr);
 	g_hash_table_destroy(basename_hash);
 }
 
-/*
- *-----------------------------------------------------------------------------
- * handling sidecars in filelist
- *-----------------------------------------------------------------------------
- */
-
-static GList *filelist_filter_out_sidecars(GList *flist)
-{
-	GList *work = flist;
-	GList *flist_filtered = nullptr;
-
-	while (work)
-		{
-		auto fd = static_cast<FileData *>(work->data);
-
-		work = work->next;
-		if (fd->parent) /* remove fd's that are children */
-			file_data_unref(fd);
-		else
-			flist_filtered = g_list_prepend(flist_filtered, fd);
-		}
-	g_list_free(flist);
-
-	return flist_filtered;
-}
-
-static void file_data_basename_hash_to_sidecars(gpointer, gpointer value, gpointer)
+void FileData::file_data_basename_hash_to_sidecars(gpointer, gpointer value, gpointer)
 {
 	auto basename_list = static_cast<GList *>(value);
 	file_data_check_sidecars(basename_list);
 }
 
-
-static gboolean is_hidden_file(const gchar *name)
-{
-	if (name[0] != '.') return FALSE;
-	if (name[1] == '\0' || (name[1] == '.' && name[2] == '\0')) return FALSE;
-	return TRUE;
-}
-
-/*
- *-----------------------------------------------------------------------------
- * the main filelist function
- *-----------------------------------------------------------------------------
- */
-
-static gboolean filelist_read_real(const gchar *dir_path, GList **files, GList **dirs, gboolean follow_symlinks)
-{
-	DIR *dp;
-	struct dirent *dir;
-	gchar *pathl;
-	GList *dlist = nullptr;
-	GList *flist = nullptr;
-	GList *xmp_files = nullptr;
-	gint (*stat_func)(const gchar *path, struct stat *buf);
-	GHashTable *basename_hash = nullptr;
-
-	g_assert(files || dirs);
-
-	if (files) *files = nullptr;
-	if (dirs) *dirs = nullptr;
-
-	pathl = path_from_utf8(dir_path);
-	if (!pathl) return FALSE;
-
-	dp = opendir(pathl);
-	if (dp == nullptr)
-		{
-		g_free(pathl);
-		return FALSE;
-		}
-
-	if (files) basename_hash = file_data_basename_hash_new();
-
-	if (follow_symlinks)
-		stat_func = stat;
-	else
-		stat_func = lstat;
-
-	while ((dir = readdir(dp)) != nullptr)
-		{
-		struct stat ent_sbuf;
-		const gchar *name = dir->d_name;
-		gchar *filepath;
-
-		if (!options->file_filter.show_hidden_files && is_hidden_file(name))
-			continue;
-
-		filepath = g_build_filename(pathl, name, NULL);
-		if (stat_func(filepath, &ent_sbuf) >= 0)
-			{
-			if (S_ISDIR(ent_sbuf.st_mode))
-				{
-				/* we ignore the .thumbnails dir for cleanliness */
-				if (dirs &&
-				    (name[0] != '.' || (name[1] != '\0' && (name[1] != '.' || name[2] != '\0'))) &&
-				    strcmp(name, GQ_CACHE_LOCAL_THUMB) != 0 &&
-				    strcmp(name, GQ_CACHE_LOCAL_METADATA) != 0 &&
-				    strcmp(name, THUMB_FOLDER_LOCAL) != 0)
-					{
-					dlist = g_list_prepend(dlist, file_data_new_local(filepath, &ent_sbuf, TRUE));
-					}
-				}
-			else
-				{
-				if (files && filter_name_exists(name))
-					{
-					FileData *fd = file_data_new_local(filepath, &ent_sbuf, FALSE);
-					flist = g_list_prepend(flist, fd);
-					if (fd->sidecar_priority && !fd->disable_grouping)
-						{
-						if (strcmp(fd->extension, ".xmp") != 0)
-							file_data_basename_hash_insert(basename_hash, fd);
-						else
-							xmp_files = g_list_append(xmp_files, fd);
-						}
-					}
-				}
-			}
-		else
-			{
-			if (errno == EOVERFLOW)
-				{
-				log_printf("stat(): EOVERFLOW, skip '%s'", filepath);
-				}
-			}
-		g_free(filepath);
-		}
-
-	closedir(dp);
-
-	g_free(pathl);
-
-	if (xmp_files)
-		{
-		g_list_foreach(xmp_files,file_data_basename_hash_insert_cb,basename_hash);
-		g_list_free(xmp_files);
-		}
-
-	if (dirs) *dirs = dlist;
-
-	if (files)
-		{
-		g_hash_table_foreach(basename_hash, file_data_basename_hash_to_sidecars, nullptr);
-
-		*files = filelist_filter_out_sidecars(flist);
-		}
-	if (basename_hash) file_data_basename_hash_free(basename_hash);
-
-	return TRUE;
-}
-
-gboolean FileData::filelist_read(FileData *dir_fd, GList **files, GList **dirs)
-{
-	return filelist_read_real(dir_fd->path, files, dirs, TRUE);
-}
-
-gboolean FileData::filelist_read_lstat(FileData *dir_fd, GList **files, GList **dirs)
-{
-	return filelist_read_real(dir_fd->path, files, dirs, FALSE);
-}
 
 FileData *FileData::file_data_new_group(const gchar *path_utf8)
 {
@@ -1545,7 +1268,7 @@ FileData *FileData::file_data_new_group(const gchar *path_utf8)
 
 	dir = remove_level_from_path(path_utf8);
 
-	filelist_read_real(dir, &files, nullptr, TRUE);
+	FileList::filelist_read_real(dir, &files, nullptr, TRUE);
 
 	fd = static_cast<FileData *>(g_hash_table_lookup(file_data_pool, path_utf8));
 	if (!fd) fd = file_data_new(path_utf8, &st, TRUE);
@@ -1557,206 +1280,6 @@ FileData *FileData::file_data_new_group(const gchar *path_utf8)
 	filelist_free(files);
 	g_free(dir);
 	return fd;
-}
-
-
-void FileData::filelist_free(GList *list)
-{
-	GList *work;
-
-	work = list;
-	while (work)
-		{
-		::file_data_unref((FileData *)work->data);
-		work = work->next;
-		}
-
-	g_list_free(list);
-}
-
-
-GList *FileData::filelist_copy(GList *list)
-{
-	GList *new_list = nullptr;
-
-	for (GList *work = list; work; work = work->next)
-		{
-		auto fd = static_cast<FileData *>(work->data);
-
-		new_list = g_list_prepend(new_list, ::file_data_ref(fd));
-		}
-
-	return g_list_reverse(new_list);
-}
-
-GList *FileData::filelist_from_path_list(GList *list)
-{
-	GList *new_list = nullptr;
-	GList *work;
-
-	work = list;
-	while (work)
-		{
-		gchar *path;
-
-		path = static_cast<gchar *>(work->data);
-		work = work->next;
-
-		new_list = g_list_prepend(new_list, file_data_new_group(path));
-		}
-
-	return g_list_reverse(new_list);
-}
-
-GList *FileData::filelist_to_path_list(GList *list)
-{
-	GList *new_list = nullptr;
-	GList *work;
-
-	work = list;
-	while (work)
-		{
-		FileData *fd;
-
-		fd = static_cast<FileData *>(work->data);
-		work = work->next;
-
-		new_list = g_list_prepend(new_list, g_strdup(fd->path));
-		}
-
-	return g_list_reverse(new_list);
-}
-
-GList *FileData::filelist_filter(GList *list, gboolean is_dir_list)
-{
-	GList *work;
-
-	if (!is_dir_list && options->file_filter.disable && options->file_filter.show_hidden_files) return list;
-
-	work = list;
-	while (work)
-		{
-		auto fd = static_cast<FileData *>(work->data);
-		const gchar *name = fd->name;
-		GList *link = work;
-		work = work->next;
-
-		if ((!options->file_filter.show_hidden_files && is_hidden_file(name)) ||
-		    (!is_dir_list && !filter_name_exists(name)) ||
-		    (is_dir_list && name[0] == '.' && (strcmp(name, GQ_CACHE_LOCAL_THUMB) == 0 ||
-						       strcmp(name, GQ_CACHE_LOCAL_METADATA) == 0)) )
-			{
-			list = g_list_remove_link(list, link);
-			::file_data_unref(fd);
-			g_list_free(link);
-			}
-		}
-
-	return list;
-}
-
-/*
- *-----------------------------------------------------------------------------
- * filelist recursive
- *-----------------------------------------------------------------------------
- */
-
-static gint filelist_sort_path_cb(gconstpointer a, gconstpointer b)
-{
-	return CASE_SORT(((FileData *)a)->path, ((FileData *)b)->path);
-}
-
-GList *FileData::filelist_sort_path(GList *list)
-{
-	return g_list_sort(list, filelist_sort_path_cb);
-}
-
-static void filelist_recursive_append(GList **list, GList *dirs)
-{
-	GList *work;
-
-	work = dirs;
-	while (work)
-		{
-		auto fd = static_cast<FileData *>(work->data);
-		GList *f;
-		GList *d;
-
-		if (filelist_read(fd, &f, &d))
-			{
-			f = filelist_filter(f, FALSE);
-			f = filelist_sort_path(f);
-			*list = g_list_concat(*list, f);
-
-			d = filelist_filter(d, TRUE);
-			d = filelist_sort_path(d);
-			filelist_recursive_append(list, d);
-			filelist_free(d);
-			}
-
-		work = work->next;
-		}
-}
-
-static void filelist_recursive_append_full(GList **list, GList *dirs, SortType method, gboolean ascend, gboolean case_sensitive)
-{
-	GList *work;
-
-	work = dirs;
-	while (work)
-		{
-		auto fd = static_cast<FileData *>(work->data);
-		GList *f;
-		GList *d;
-
-		if (filelist_read(fd, &f, &d))
-			{
-			f = filelist_filter(f, FALSE);
-			f = filelist_sort_full(f, method, ascend, case_sensitive, reinterpret_cast<GCompareFunc>(filelist_sort_file_cb));
-			*list = g_list_concat(*list, f);
-
-			d = filelist_filter(d, TRUE);
-			d = filelist_sort_path(d);
-			filelist_recursive_append_full(list, d, method, ascend, case_sensitive);
-			filelist_free(d);
-			}
-
-		work = work->next;
-		}
-}
-
-GList *FileData::filelist_recursive(FileData *dir_fd)
-{
-	GList *list;
-	GList *d;
-
-	if (!::filelist_read(dir_fd, &list, &d)) return nullptr;
-	list = filelist_filter(list, FALSE);
-	list = filelist_sort_path(list);
-
-	d = filelist_filter(d, TRUE);
-	d = filelist_sort_path(d);
-	filelist_recursive_append(&list, d);
-	filelist_free(d);
-
-	return list;
-}
-
-GList *FileData::filelist_recursive_full(FileData *dir_fd, SortType method, gboolean ascend, gboolean case_sensitive)
-{
-	GList *list;
-	GList *d;
-
-	if (!::filelist_read(dir_fd, &list, &d)) return nullptr;
-	list = filelist_filter(list, FALSE);
-	list = filelist_sort_full(list, method, ascend, case_sensitive, reinterpret_cast<GCompareFunc>(filelist_sort_file_cb));
-
-	d = filelist_filter(d, TRUE);
-	d = filelist_sort_path(d);
-	filelist_recursive_append_full(&list, d, method, ascend, case_sensitive);
-	filelist_free(d);
-
-	return list;
 }
 
 /*
